requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
distro==1.9.0
Django==4.2.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
openai==1.61.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
